This commit was manufactured by cvs2svn to create tag 'release_grass5beta7_20_april_2000'.
git-svn-id: https://svn.osgeo.org/grass/grass/tags/release_20000420_grass_5_0_0beta7@592 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ src/raster/r.drain/cmd/drain_cmd.h
@@ -1,4 +1,4 @@
 int drain_path_finder(POINT *);
-POINT *make_neighbors_list(POINT *, int, int, int, int, int, int);
+POINT *make_neighbors_list(POINT *, int, int, double, int, int, double);
 POINT *make_point(POINT *, int, int, double);
 int free_list(POINT *);
